+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 
 # 1.7.08 (2022-03-04)
 
+## UI:
+
+- "Render Shot Prefix" parameter was renamed Render Sequence Prefix
+
 ## Code:
 
 - Cleaned the debug folder to make it a well-integrated package
-
+- render_sequence_prefix renamed to render_sequence_prefix
 
 # 1.7.1 (2022-02-11)