+++ CHANGELOG.md
@@ -22,16 +22,24 @@
 ## Grease Pencil:
 - Fixed bug when painting on hidden objects
 
+<<<<<<< HEAD
 ## UI:
 - Added button to take range
 - Added information for output render
 - Added the main panel items menu in the Render Panel
+=======
+fixed res ouput for sep video and for playblast
+fixed shots play mode issues on version 3
+>>>>>>> Fix: Shots Play Mode issues on Blender V 3.x
 
 ## Code:
 - Code cleaning to match Flake8 rules
 - Refactor code for output media
 ***Warning: temp directory names have changed ***
 
+
+Debug
+Added debug function to fix entities parent in old blender files
 
 Debug
 Added debug function to fix entities parent in old blender files