+++ CHANGELOG.md
@@ -42,8 +42,13 @@
 - Added a checkbox in the render panel to choose to keep the intermediate rendered images
 - Added a Reset Render Properties button
 
+<<<<<<< HEAD
 ## Grease Pencil:
 - Fixed bug when painting on hidden objects
+=======
+fixed res ouput for sep video and for playblast
+fixed shots play mode issues on version 3
+>>>>>>> Fix: Shots Play Mode issues on Blender V 3.x
 
 ## UI:
 
@@ -59,6 +64,9 @@
 
 ## Debug
 - Added debug function to fix entities parent in old blender files
+
+Debug
+Added debug function to fix entities parent in old blender files
 
 Debug
 Added debug function to fix entities parent in old blender files